파일 복구

+++ spring-breeze-erp-v3/back/bootSBerp/src/main/java/com/sb/erp/emp/dto/WelcomeMailTargetDto.java
@@ -11,7 +11,7 @@ import lombok.Data;
 
 @Data
 public class WelcomeMailTargetDto {
-    private int    empId;
+    private long   empId;
     private String empName;
     private String empEmail;
     private String comName;
